scrapping/requirements.txt
# Web Scraping Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
selenium>=4.15.0
openpyxl>=3.1.0

# Backend Dependencies
fastapi
uvicorn
python-dotenv

# Additional utilities
lxml>=4.9.0

# Forecasting Dependencies
prophet>=1.1.0
torch>=2.0.0
chronos-forecasting
numpy>=1.24.0
scikit-learn>=1.3.0
